ISPN-13752 - Remote LDAP Kerberos support
