+++ server/tests/src/test/java/org/infinispan/server/test/core/LdapServerListener.java
@@ -1,19 +1,5 @@
 package org.infinispan.server.test.core;
 
-import java.io.File;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-
-import javax.security.auth.kerberos.KerberosPrincipal;
-
-import org.apache.directory.server.kerberos.shared.crypto.encryption.KerberosKeyFactory;
-import org.apache.directory.server.kerberos.shared.keytab.Keytab;
-import org.apache.directory.server.kerberos.shared.keytab.KeytabEntry;
-import org.apache.directory.shared.kerberos.KerberosTime;
-import org.apache.directory.shared.kerberos.codec.types.EncryptionType;
-import org.apache.directory.shared.kerberos.components.EncryptionKey;
 import org.infinispan.commons.test.Exceptions;
 import org.infinispan.commons.test.ThreadLeakChecker;
 import org.infinispan.server.test.core.ldap.AbstractLdapServer;
@@ -29,25 +15,18 @@ public class LdapServerListener implements InfinispanServerListener {
    private static final String DEFAULT_LDIF = "ldif/infinispan.ldif";
    private static final String KERBEROS_LDIF = "ldif/infinispan-kerberos.ldif";
 
-   private final String initLDIF;
    private AbstractLdapServer ldapServer;
-   private boolean withKdc;
 
    public LdapServerListener() {
       this(false);
    }
 
    public LdapServerListener(boolean withKdc) {
-      if (withKdc) {
-         initLDIF = KERBEROS_LDIF;
-      } else {
-         initLDIF = DEFAULT_LDIF;
-      }
-      this.withKdc = withKdc;
       if ("apache".equals(LDAP_SERVER)) {
-         ldapServer = new ApacheLdapServer();
+         ldapServer = new ApacheLdapServer(withKdc, withKdc ? KERBEROS_LDIF : DEFAULT_LDIF);
+      // when using the remote ldap server, you should overwrite the content of the ldif files before running the test
       } else if ("remote".equals(LDAP_SERVER)) {
-         ldapServer = new RemoteLdapServer();
+         ldapServer = new RemoteLdapServer(withKdc ? KERBEROS_LDIF : DEFAULT_LDIF);
       } else {
          throw new IllegalStateException("Unsupported LDAP Server: " + LDAP_SERVER);
       }
@@ -56,14 +35,7 @@ public class LdapServerListener implements InfinispanServerListener {
    @Override
    public void before(InfinispanServerDriver driver) {
       Exceptions.unchecked(() -> {
-         if (withKdc) {
-            generateKeyTab(new File(driver.getConfDir(), "hotrod.keytab"), "hotrod/datagrid@INFINISPAN.ORG", "hotrodPassword");
-            generateKeyTab(new File(driver.getConfDir(), "http.keytab"), "HTTP/localhost@INFINISPAN.ORG", "httpPassword");
-         }
-         ldapServer.start(driver.getCertificateFile("server").getAbsolutePath(), this.initLDIF);
-         if (withKdc) {
-            ldapServer.startKdc();
-         }
+         ldapServer.start(driver.getCertificateFile("server").getAbsolutePath(), driver.getConfDir());
       });
    }
 
@@ -79,30 +51,5 @@ public class LdapServerListener implements InfinispanServerListener {
       ThreadLeakChecker.ignoreThreadsContaining("pool-.*thread-");
       //
       ThreadLeakChecker.ignoreThreadsContaining("^Thread-\\d+$");
-   }
-
-   public static String generateKeyTab(File keyTabFile, String... credentials) {
-      List<KeytabEntry> entries = new ArrayList<>();
-      KerberosTime ktm = new KerberosTime();
-
-      for (int i = 0; i < credentials.length; ) {
-         String principal = credentials[i++];
-         String password = credentials[i++];
-
-         for (Map.Entry<EncryptionType, EncryptionKey> keyEntry : KerberosKeyFactory.getKerberosKeys(principal, password)
-               .entrySet()) {
-            EncryptionKey key = keyEntry.getValue();
-            entries.add(new KeytabEntry(principal, KerberosPrincipal.KRB_NT_PRINCIPAL, ktm, (byte) key.getKeyVersion(), key));
-         }
-      }
-
-      Keytab keyTab = Keytab.getInstance();
-      keyTab.setEntries(entries);
-      try {
-         keyTab.write(keyTabFile);
-         return keyTabFile.getAbsolutePath();
-      } catch (IOException e) {
-         throw new IllegalStateException("Cannot create keytab: " + keyTabFile, e);
-      }
    }
 }
