+++ server/tests/src/test/java/org/infinispan/server/test/core/ldap/AbstractLdapServer.java
@@ -1,8 +1,6 @@
 package org.infinispan.server.test.core.ldap;
 
-import java.io.IOException;
-
-import org.apache.directory.api.ldap.model.exception.LdapInvalidDnException;
+import java.io.File;
 
 public abstract class AbstractLdapServer {
 
@@ -10,9 +8,7 @@ public abstract class AbstractLdapServer {
    public static final String TEST_LDAP_PRINCIPAL = "org.infinispan.test.ldap.principal";
    public static final String TEST_LDAP_CREDENTIAL = "org.infinispan.test.ldap.credential";
 
-   public abstract void start(String keystoreFile, String initLDIF) throws Exception;
+   public abstract void start(String keystoreFile, File confDir) throws Exception;
 
    public abstract void stop() throws Exception;
-
-   public abstract void startKdc() throws IOException, LdapInvalidDnException;
 }
