+++ server/tests/src/test/java/org/infinispan/server/test/core/ldap/ApacheLdapServer.java
@@ -1,7 +1,13 @@
 package org.infinispan.server.test.core.ldap;
 
+import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+
+import javax.security.auth.kerberos.KerberosPrincipal;
 
 import org.apache.directory.api.ldap.model.entry.DefaultEntry;
 import org.apache.directory.api.ldap.model.exception.LdapInvalidDnException;
@@ -21,10 +27,16 @@ import org.apache.directory.server.core.kerberos.KeyDerivationInterceptor;
 import org.apache.directory.server.factory.ServerAnnotationProcessor;
 import org.apache.directory.server.kerberos.KerberosConfig;
 import org.apache.directory.server.kerberos.kdc.KdcServer;
+import org.apache.directory.server.kerberos.shared.crypto.encryption.KerberosKeyFactory;
+import org.apache.directory.server.kerberos.shared.keytab.Keytab;
+import org.apache.directory.server.kerberos.shared.keytab.KeytabEntry;
 import org.apache.directory.server.ldap.LdapServer;
 import org.apache.directory.server.protocol.shared.transport.TcpTransport;
 import org.apache.directory.server.protocol.shared.transport.Transport;
 import org.apache.directory.server.protocol.shared.transport.UdpTransport;
+import org.apache.directory.shared.kerberos.KerberosTime;
+import org.apache.directory.shared.kerberos.codec.types.EncryptionType;
+import org.apache.directory.shared.kerberos.components.EncryptionKey;
 import org.infinispan.commons.util.Util;
 import org.infinispan.server.test.core.AbstractInfinispanServerDriver;
 
@@ -40,8 +52,12 @@ public class ApacheLdapServer extends AbstractLdapServer {
    private DirectoryService directoryService;
    private LdapServer ldapServer;
    private KdcServer kdcServer;
+   private boolean withKdc;
+   private String initLDIF;
 
-   public ApacheLdapServer() {
+   public ApacheLdapServer(boolean withKdc, String initLDIF) {
+      this.withKdc = withKdc;
+      this.initLDIF = initLDIF;
    }
 
    @CreateLdapServer(transports = {@CreateTransport(protocol = "LDAP", port = LDAP_PORT, address = LDAP_HOST)})
@@ -93,10 +109,17 @@ public class ApacheLdapServer extends AbstractLdapServer {
    }
 
    @Override
-   public void start(String keystoreFile, String initLDIF) throws Exception {
+   public void start(String keystoreFile, File confDir) throws Exception {
+      if (withKdc) {
+         generateKeyTab(new File(confDir, "hotrod.keytab"), "hotrod/datagrid@INFINISPAN.ORG", "hotrodPassword");
+         generateKeyTab(new File(confDir, "http.keytab"), "HTTP/localhost@INFINISPAN.ORG", "httpPassword");
+      }
       createDs();
       createLdap(keystoreFile, initLDIF);
       ldapServer.start();
+      if (withKdc) {
+         startKdc();
+      }
    }
 
    @Override
@@ -113,8 +136,7 @@ public class ApacheLdapServer extends AbstractLdapServer {
       }
    }
 
-   @Override
-   public void startKdc() throws IOException, LdapInvalidDnException {
+   private void startKdc() throws IOException, LdapInvalidDnException {
       createKdc();
       kdcServer.start();
    }
@@ -136,5 +158,30 @@ public class ApacheLdapServer extends AbstractLdapServer {
 
       kdcServer.setDirectoryService(directoryService);
       this.kdcServer = kdcServer;
+   }
+
+   public static String generateKeyTab(File keyTabFile, String... credentials) {
+      List<KeytabEntry> entries = new ArrayList<>();
+      KerberosTime ktm = new KerberosTime();
+
+      for (int i = 0; i < credentials.length; ) {
+         String principal = credentials[i++];
+         String password = credentials[i++];
+
+         for (Map.Entry<EncryptionType, EncryptionKey> keyEntry : KerberosKeyFactory.getKerberosKeys(principal, password)
+               .entrySet()) {
+            EncryptionKey key = keyEntry.getValue();
+            entries.add(new KeytabEntry(principal, KerberosPrincipal.KRB_NT_PRINCIPAL, ktm, (byte) key.getKeyVersion(), key));
+         }
+      }
+
+      Keytab keyTab = Keytab.getInstance();
+      keyTab.setEntries(entries);
+      try {
+         keyTab.write(keyTabFile);
+         return keyTabFile.getAbsolutePath();
+      } catch (IOException e) {
+         throw new IllegalStateException("Cannot create keytab: " + keyTabFile, e);
+      }
    }
 }
